+++ Software/Firmware/Src/ILI9488/ILI9488_STM32_Driver.h
@@ -63,9 +63,8 @@ void ILI9488_Init(void);
 void ILI9488_Fill_Screen(uint16_t Colour);
 void ILI9488_Draw_Pixel(uint16_t X, uint16_t Y, uint16_t Colour);
 void ILI9488_Draw_Colour_Burst(uint16_t Colour, uint32_t Size);
-// TODO rewrite following functions for 16 bit colour
-void ILI9488_Draw_Rectangle(uint16_t X, uint16_t Y, uint16_t Width, uint16_t Height, uint8_t Colour);
-void ILI9488_Draw_Horizontal_Line(uint16_t X, uint16_t Y, uint16_t Width, uint8_t Colour);
-void ILI9488_Draw_Vertical_Line(uint16_t X, uint16_t Y, uint16_t Height, uint8_t Colour);
+void ILI9488_Draw_Rectangle(uint16_t X, uint16_t Y, uint16_t Width, uint16_t Height, uint16_t Colour);
+void ILI9488_Draw_Horizontal_Line(uint16_t X, uint16_t Y, uint16_t Width, uint16_t Colour);
+void ILI9488_Draw_Vertical_Line(uint16_t X, uint16_t Y, uint16_t Height, uint16_t Colour);
 	
 #endif
